requirements.txt
anthropic>=0.39.0
kg-gen>=0.1.6
spacy>=3.7.0
sentence-transformers>=2.7.0
scikit-learn>=1.3.0
transformers>=4.40.0
torch>=2.0.0
networkx>=3.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=8.0.0
